Make catch_warning_msg a utility; use specialized instance forPandas block warning
